requirements.txt
itsdangerous==2.0.1
Pillow==9.1.1
pycryptodome==3.14.1
PyQt6==6.3.0
sympy==1.8
